added inputComponent to operator
